Update LICENSE to AGPL-3.0 to align with YOLOv8
This PR updates the license of the FastSAM repository from its current state to AGPL-3.0. This change is to ensure consistency and compliance with the Ultralytics YOLOv8 license, which FastSAM utilizes for object detection tasks.

As FastSAM integrates YOLOv8 for certain functionalities, it's essential that our licensing reflects the compliance requirements of the open-source components we use. The AGPL-3.0 license is a strong copyleft license that mandates that any modifications to AGPL-3.0 licensed code be made available to the public when distributed, under the same license. This aligns with our commitment to open-source values and ensures that enhancements in FastSAM that are based on YOLOv8's technology are shared with the community.

We believe this step is important for maintaining the legal and ethical use of open-source software. Thank you for your prompt attention to this matter.